[Bronze V] Title: Multiply, Time: 124 ms, Memory: 14072 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.java
@@ -3,8 +3,8 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         br.readLine();
-        int x = Integer.parseInt(br.readLine());
-        int y = Integer.parseInt(br.readLine());
+        long x = Long.parseLong(br.readLine());
+        long y = Long.parseLong(br.readLine());
         System.out.println(x*y);
     }
 }
